+++ examples/example1.php
@@ -18,8 +18,10 @@ use Eluceo\iCal\Domain\Entity\Event;
 use Eluceo\iCal\Domain\ValueObject\Alarm;
 use Eluceo\iCal\Domain\ValueObject\Attachment;
 use Eluceo\iCal\Domain\ValueObject\DateTime;
+use Eluceo\iCal\Domain\ValueObject\EmailAddress;
 use Eluceo\iCal\Domain\ValueObject\GeographicPosition;
 use Eluceo\iCal\Domain\ValueObject\Location;
+use Eluceo\iCal\Domain\ValueObject\Organizer;
 use Eluceo\iCal\Domain\ValueObject\TimeSpan;
 use Eluceo\iCal\Domain\ValueObject\Uri;
 use Eluceo\iCal\Presentation\Factory\CalendarFactory;
@@ -31,6 +33,10 @@ $event = new Event();
 $event
     ->setSummary('Christmas Eve')
     ->setDescription('Lorem Ipsum Dolor...')
+    ->setOrganizer(new Organizer(
+        new EmailAddress('john.doe@example.com'),
+        'John Doe'
+    ))
     ->setLocation(
         (new Location('Neuschwansteinstraße 20, 87645 Schwangau', 'Schloss Neuschwanstein'))
             ->withGeographicPosition(new GeographicPosition(47.557579, 10.749704))