+++ src/Domain/Entity/Event.php
@@ -15,6 +15,7 @@ use Eluceo\iCal\Domain\ValueObject\Alarm;
 use Eluceo\iCal\Domain\ValueObject\Attachment;
 use Eluceo\iCal\Domain\ValueObject\Location;
 use Eluceo\iCal\Domain\ValueObject\Occurrence;
+use Eluceo\iCal\Domain\ValueObject\Organizer;
 use Eluceo\iCal\Domain\ValueObject\Timestamp;
 use Eluceo\iCal\Domain\ValueObject\UniqueIdentifier;
 
@@ -26,6 +27,7 @@ class Event
     private ?string $description = null;
     private ?Occurrence $occurrence = null;
     private ?Location $location = null;
+    private ?Organizer $organizer = null;
 
     /**
      * @var array<Alarm>
@@ -148,6 +150,25 @@ class Event
     public function hasLocation(): bool
     {
         return $this->location !== null;
+    }
+
+    public function getOrganizer(): Organizer
+    {
+        assert($this->organizer !== null);
+
+        return $this->organizer;
+    }
+
+    public function setOrganizer(?Organizer $organizer): self
+    {
+        $this->organizer = $organizer;
+
+        return $this;
+    }
+
+    public function hasOrganizer(): bool
+    {
+        return $this->organizer !== null;
     }
 
     /**